+++ ubo_app/services/050-vscode/reducer.py
@@ -3,8 +3,20 @@ from __future__ import annotations
 
 from dataclasses import replace
 
-from redux import InitAction, InitializationActionError
+from redux import (
+    BaseEvent,
+    CompleteReducerResult,
+    InitAction,
+    InitializationActionError,
+)
+from ubo_gui.constants import SUCCESS_COLOR
 
+from ubo_app.store.services.notifications import (
+    Importance,
+    Notification,
+    NotificationDisplayType,
+    NotificationsAddAction,
+)
 from ubo_app.store.services.vscode import (
     VSCodeAction,
     VSCodeDoneDownloadingAction,
@@ -15,7 +27,12 @@ from ubo_app.store.services.vscode import (
 )
 
 
-def reducer(state: VSCodeState | None, action: VSCodeAction) -> VSCodeState:
+def reducer(
+    state: VSCodeState | None,
+    action: VSCodeAction,
+) -> (
+    CompleteReducerResult[VSCodeState, NotificationsAddAction, BaseEvent] | VSCodeState
+):
     if state is None:
         if isinstance(action, InitAction):
             return VSCodeState()
@@ -31,12 +48,34 @@ def reducer(state: VSCodeState | None, action: VSCodeAction) -> VSCodeState:
         return replace(state, is_pending=True)
 
     if isinstance(action, VSCodeSetStatusAction):
-        return replace(
+        actions = []
+        if (
+            state.status
+            and not state.status.is_running
+            and action.status
+            and action.status.is_running
+        ):
+            actions.append(
+                NotificationsAddAction(
+                    notification=Notification(
+                        title='VSCode',
+                        content='Service is running',
+                        icon='󰨞',
+                        importance=Importance.MEDIUM,
+                        color=SUCCESS_COLOR,
+                        display_type=NotificationDisplayType.FLASH,
+                    ),
+                ),
+            )
+
+        state = replace(
             state,
             is_pending=False,
             is_binary_installed=action.is_binary_installed,
             is_logged_in=action.is_logged_in,
             status=action.status,
         )
+
+        return CompleteReducerResult(state=state, actions=actions)
 
     return state